fix: Github detector should not use firebase
remove usage of firebase from github detector
 - side-effect: downloading a github url won't auto download latest tag and it will be required to pass it explicitly
