+++ internal/getter/hub.go
@@ -1,10 +1,13 @@
 package getter
 
 import (
+	"context"
 	"fmt"
 	"net/http"
 	"net/url"
 	"strings"
+
+	"github.com/cloudquery/cloudquery/internal/firebase"
 )
 
 type HubDetector struct {
@@ -57,4 +60,26 @@ func (HubDetector) detectHTTP(src string) (string, bool, error) {
 	}
 
 	return "git::" + _url.String(), true, nil
+}
+
+func addLatestTag(_url *url.URL, owner, repo string) error {
+	client := firebase.New(firebase.CloudQueryRegistryURL)
+	org, ok := repoToFirebasePath[owner]
+	if !ok {
+		org = owner
+	}
+	latest, err := client.GetLatestPolicyRelease(context.Background(), org, repo)
+
+	if err != nil {
+		return fmt.Errorf("failed to find latest version: %w", err)
+	}
+
+	if latest == "" {
+		return nil
+	}
+
+	q := _url.Query()
+	q.Add("ref", latest)
+	_url.RawQuery = q.Encode()
+	return nil
 }
